+++ packages/search/global-search/src/__tests__/unit/GlobalQuickSearch.test.tsx
@@ -6,6 +6,7 @@ import GlobalQuickSearchWithAnalytics, {
 } from '../../components/GlobalQuickSearch';
 import * as AnalyticsHelper from '../../util/analytics-event-helper';
 import { CreateAnalyticsEventFn } from '../../components/analytics/types';
+import { ReferralContextIdentifiers } from '../../components/GlobalQuickSearchWrapper';
 
 const noop = () => {};
 const DEFAULT_PROPS = {
@@ -54,7 +55,23 @@ describe('GlobalQuickSearch', () => {
       .prop('onSearchInput');
     onSearchInput({ target: { value: 'foo' } });
 
-    expect(searchMock).toHaveBeenCalledWith('foo');
+    expect(searchMock).toHaveBeenCalledWith('foo', 0);
+  });
+
+  it('should fire searches with the queryVersion parameter incrementing', () => {
+    const searchMock = jest.fn();
+    const wrapper = render({ onSearch: searchMock });
+
+    const onSearchInput: Function = wrapper
+      .children()
+      .first()
+      .prop('onSearchInput');
+
+    onSearchInput({ target: { value: 'foo' } });
+    expect(searchMock).toHaveBeenNthCalledWith(1, 'foo', 0);
+
+    onSearchInput({ target: { value: 'foo' } });
+    expect(searchMock).toHaveBeenNthCalledWith(2, 'foo', 1);
   });
 
   it('should trim the search input', () => {
@@ -67,7 +84,7 @@ describe('GlobalQuickSearch', () => {
       .prop('onSearchInput');
     onSearchInput({ target: { value: '  pattio   ' } });
 
-    expect(searchMock).toHaveBeenCalledWith('pattio');
+    expect(searchMock).toHaveBeenCalledWith('pattio', 0);
   });
 
   describe('Search result events', () => {
@@ -76,6 +93,7 @@ describe('GlobalQuickSearch', () => {
       (
         eventData: AnalyticsHelper.KeyboardControlEvent,
         searchSessionId: string,
+        referralContextIdentifiers?: ReferralContextIdentifiers,
         createAnalyticsEvent?: CreateAnalyticsEventFn | undefined,
       ) => void
     >;
@@ -83,6 +101,7 @@ describe('GlobalQuickSearch', () => {
       (
         eventData: AnalyticsHelper.SelectedSearchResultEvent,
         searchSessionId: string,
+        referralContextIdentifiers?: ReferralContextIdentifiers,
         createAnalyticsEvent?: CreateAnalyticsEventFn | undefined,
       ) => void
     >;
@@ -90,6 +109,7 @@ describe('GlobalQuickSearch', () => {
       (
         eventData: AnalyticsHelper.AdvancedSearchSelectedEvent,
         searchSessionId: string,
+        referralContextIdentifiers?: ReferralContextIdentifiers,
         createAnalyticsEvent?: CreateAnalyticsEventFn | undefined,
       ) => void
     >;