+++ packages/search/global-search/src/__tests__/unit/common/ContainerSearchResults.test.tsx
@@ -114,8 +114,6 @@ const assertJiraNoRecentActivity = (element: JSX.Element) => {
     props: {
       query: 'query',
       analyticsData: { resultsCount: 0, wasOnNoResultsScreen: true },
-      showKeyboardLozenge: false,
-      showSearchIcon: false,
     },
   });
 };
@@ -147,8 +145,6 @@ const assertJiraAdvancedSearchGroup = (element: JSX.Element) => {
     props: {
       analyticsData: { resultsCount: 10 },
       query: 'query',
-      showKeyboardLozenge: true,
-      showSearchIcon: true,
     },
   });
 };
@@ -258,6 +254,20 @@ const getJiraPostQueryResults = () => [
     items: issues,
     key: 'issues',
     title: messages.jira_search_result_issues_heading,
+  },
+  {
+    items: [
+      {
+        analyticsType: 'link-postquery-advanced-search-jira',
+        contentType: 'jira-issue',
+        href: 'jiraUrl',
+        name: 'jira',
+        resultId: 'search-jira',
+        resultType: 'JiraIssueAdvancedSearch',
+      },
+    ],
+    key: 'issue-advanced',
+    title: undefined,
   },
   {
     items: boards,
@@ -384,6 +394,7 @@ const getPreqQueryResults = (product: QuickSearchContext) =>
       });
 
       it('should return postQueryGroups', () => {
+        getAdvancedSearchUrlSpy.mockReturnValue('jiraUrl');
         const { getPostQueryGroups } = getProps();
         const postQueryGroups = getPostQueryGroups();
         expect(postQueryGroups).toMatchObject(getPostQueryResults(product));
@@ -409,8 +420,6 @@ describe('jira', () => {
       props: {
         analyticsData: { resultsCount: 10 },
         query: '',
-        showKeyboardLozenge: false,
-        showSearchIcon: true,
       },
     });
   });