+++ packages/search/global-search/src/__tests__/unit/confluence/ConfluenceQuickSearchContainer.test.tsx
@@ -91,6 +91,33 @@ describe('ConfluenceQuickSearchContainer', () => {
     });
   });
 
+  it('should call cross product search client with correct query version', async () => {
+    const searchSpy = jest.spyOn(noResultsCrossProductSearchClient, 'search');
+    const dummyQueryVersion = 123;
+
+    const wrapper = render({
+      confluenceClient: noResultsConfluenceClient,
+      crossProductSearchClient: noResultsCrossProductSearchClient,
+    });
+
+    const quickSearchContainer = wrapper.find(QuickSearchContainer);
+    (quickSearchContainer.props() as QuickSearchContainerProps).getSearchResults(
+      'query',
+      sessionId,
+      100,
+      dummyQueryVersion,
+    );
+
+    expect(searchSpy).toHaveBeenCalledWith(
+      'query',
+      expect.any(Object),
+      expect.any(Array),
+      dummyQueryVersion,
+    );
+
+    searchSpy.mockRestore();
+  });
+
   it('should return ab test data', async () => {
     const abTest: ABTest = {
       abTestId: 'abTestId',
@@ -134,6 +161,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults).toMatchObject({
@@ -160,7 +188,7 @@ describe('ConfluenceQuickSearchContainer', () => {
     });
   });
 
-  it('should use CPUs for people results when enabled', async () => {
+  it('should use CPUS for people results when enabled', async () => {
     const wrapper = render({
       useCPUSForPeopleResults: true,
       crossProductSearchClient: {
@@ -188,6 +216,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults.results.people).toEqual([
@@ -216,6 +245,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults.results.people).toEqual([