+++ packages/search/global-search/src/components/ResultList.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
+import SearchIcon from '@atlaskit/icon/glyph/search';
 import {
   ObjectResult as ObjectResultComponent,
   PersonResult as PersonResultComponent,
   ContainerResult as ContainerResultComponent,
+  ResultBase,
 } from '@atlaskit/quick-search';
 import { FormattedMessage } from 'react-intl';
 import { messages } from '../messages';
@@ -15,11 +17,13 @@ import {
   ConfluenceObjectResult,
   JiraProjectType,
   ContentType,
+  AnalyticsType,
 } from '../model/Result';
 import { SelectedIcon } from './styled';
 import { getAvatarForConfluenceObjectResult } from '../util/confluence-avatar-util';
 import { getDefaultAvatar } from '../util/jira-avatar-util';
 import DarkReturn from '../assets/DarkReturn';
+import Return from '../assets/Return';
 
 export interface Props {
   results: Result[];
@@ -199,6 +203,20 @@ export default class ResultList extends React.Component<Props> {
               analyticsData={analyticsData}
               selectedIcon={selectedIcon}
               target="_blank"
+            />
+          );
+        }
+
+        case ResultType.JiraIssueAdvancedSearch: {
+          return (
+            <ResultBase
+              href={result.href}
+              resultId="jira-advanced-issue-search"
+              text={<FormattedMessage {...messages.jira_view_all_issues} />}
+              icon={<SearchIcon size="medium" label="View all issues" />}
+              type={AnalyticsType.LinkPostQueryAdvancedSearchJira}
+              key={uniqueKey}
+              elemAfter={<Return />}
             />
           );
         }