+++ packages/search/global-search/src/components/common/QuickSearchContainer.tsx
@@ -1,6 +1,10 @@
 import * as React from 'react';
 import uuid from 'uuid/v4';
-import { LinkComponent, Logger } from '../GlobalQuickSearchWrapper';
+import {
+  LinkComponent,
+  Logger,
+  ReferralContextIdentifiers,
+} from '../GlobalQuickSearchWrapper';
 import GlobalQuickSearch from '../GlobalQuickSearch';
 import performanceNow from '../../util/performance-now';
 import {
@@ -39,9 +43,11 @@ export interface Props {
     query: string,
     sessionId: string,
     startTime: number,
+    queryVersion: number,
   ): Promise<ResultsWithTiming>;
   getAbTestData(sessionId: string): Promise<ABTest>;
   getAutocomplete?(query: string): Promise<string[]>;
+  referralContextIdentifiers?: ReferralContextIdentifiers;
 
   /**
    * return displayed groups from result groups
@@ -90,6 +96,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
 
   // used to terminate if component is unmounted while waiting for a promise
   unmounted: boolean = false;
+  latestQueryVersion: number = 0;
 
   constructor(props: Props) {
     super(props);
@@ -129,8 +136,9 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     this.unmounted = true;
   }
 
-  doSearch = async (query: string) => {
+  doSearch = async (query: string, queryVersion: number) => {
     const startTime: number = performanceNow();
+    this.latestQueryVersion = queryVersion;
 
     this.setState({
       isLoading: true,
@@ -141,6 +149,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
         query,
         this.state.searchSessionId,
         startTime,
+        queryVersion,
       );
 
       if (this.unmounted) {
@@ -241,6 +250,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       createAnalyticsEvent,
       getDisplayedResults,
       enablePreQueryFromAggregator,
+      referralContextIdentifiers,
     } = this.props;
     if (createAnalyticsEvent && getDisplayedResults) {
       const elapsedMs: number = requestStartTime
@@ -265,6 +275,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
         searchSessionId,
         createAnalyticsEvent,
         abTest,
+        referralContextIdentifiers,
         experimentRequestDurationMs,
         !!enablePreQueryFromAggregator,
       );
@@ -286,7 +297,11 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       ...timings,
     };
 
-    const { createAnalyticsEvent, getDisplayedResults } = this.props;
+    const {
+      createAnalyticsEvent,
+      getDisplayedResults,
+      referralContextIdentifiers,
+    } = this.props;
     if (createAnalyticsEvent && getDisplayedResults) {
       const resultsArray: Result[][] = resultMapToArray(
         getDisplayedResults(searchResults, abTest),
@@ -301,11 +316,12 @@ export class QuickSearchContainer extends React.Component<Props, State> {
         latestSearchQuery,
         createAnalyticsEvent,
         abTest,
+        referralContextIdentifiers,
       );
     }
   };
 
-  handleSearch = (newLatestSearchQuery: string) => {
+  handleSearch = (newLatestSearchQuery: string, queryVersion: number) => {
     if (this.state.latestSearchQuery !== newLatestSearchQuery) {
       this.setState({
         latestSearchQuery: newLatestSearchQuery,
@@ -331,12 +347,12 @@ export class QuickSearchContainer extends React.Component<Props, State> {
           ),
       );
     } else {
-      this.doSearch(newLatestSearchQuery);
+      this.doSearch(newLatestSearchQuery, queryVersion);
     }
   };
 
   retrySearch = () => {
-    this.handleSearch(this.state.latestSearchQuery);
+    this.handleSearch(this.state.latestSearchQuery, this.latestQueryVersion);
   };
 
   handleMount = async () => {