+++ packages/search/global-search/src/components/jira/JiraQuickSearchContainer.tsx
@@ -7,6 +7,7 @@ import {
 import styled from 'styled-components';
 import { gridSize } from '@atlaskit/theme';
 import { withAnalytics } from '@atlaskit/analytics';
+import { withAnalyticsEvents } from '@atlaskit/analytics-next';
 import { CancelableEvent } from '@atlaskit/quick-search';
 import StickyFooter from '../common/StickyFooter';
 import { CreateAnalyticsEventFn } from '../analytics/types';
@@ -36,6 +37,7 @@ import {
   handlePromiseError,
   JiraEntityTypes,
   redirectToJiraAdvancedSearch,
+  ADVANCED_JIRA_SEARCH_RESULT_ID,
 } from '../SearchResultsUtil';
 import {
   ContentType,
@@ -44,6 +46,7 @@ import {
   ResultsWithTiming,
   GenericResultMap,
   JiraResultsMap,
+  AnalyticsType,
 } from '../../model/Result';
 import { getUniqueResultId } from '../ResultList';
 import {
@@ -52,6 +55,10 @@ import {
   ABTest,
 } from '../../api/CrossProductSearchClient';
 import performanceNow from '../../util/performance-now';
+import {
+  fireSelectedAdvancedSearch,
+  AdvancedSearchSelectedEvent,
+} from '../../util/analytics-event-helper';
 import AdvancedIssueSearchLink from './AdvancedIssueSearchLink';
 
 const JIRA_RESULT_LIMIT = 6;
@@ -147,6 +154,37 @@ export class JiraQuickSearchContainer extends React.Component<
     }
   };
 
+  handleAdvancedSearch = (
+    event: CancelableEvent,
+    entity: string,
+    query: string,
+    searchSessionId: string,
+    analyticsData: Object,
+    isLoading: boolean,
+  ) => {
+    const {
+      referralContextIdentifiers,
+      onAdvancedSearch = () => {},
+    } = this.props;
+    const eventData = {
+      resultId: ADVANCED_JIRA_SEARCH_RESULT_ID,
+      ...analyticsData,
+      query,
+      // queryversion is missing
+      contentType: entity,
+      type: AnalyticsType.AdvancedSearchJira,
+      isLoading,
+    } as AdvancedSearchSelectedEvent;
+
+    fireSelectedAdvancedSearch(
+      eventData,
+      searchSessionId,
+      referralContextIdentifiers,
+      this.props.createAnalyticsEvent,
+    );
+    onAdvancedSearch(event, entity, query, searchSessionId);
+  };
+
   getSearchResultsComponent = ({
     retrySearch,
     latestSearchQuery,
@@ -158,7 +196,6 @@ export class JiraQuickSearchContainer extends React.Component<
     searchSessionId,
   }: SearchResultProps) => {
     const query = latestSearchQuery;
-    const isPreQuery = !query; // it's true if the query is empty
     const {
       referralContextIdentifiers,
       onAdvancedSearch = () => {},
@@ -184,7 +221,14 @@ export class JiraQuickSearchContainer extends React.Component<
                 query={query}
                 analyticsData={{ resultsCount: 0, wasOnNoResultsScreen: true }}
                 onClick={(mouseEvent, entity) =>
-                  onAdvancedSearch(mouseEvent, entity, query, searchSessionId)
+                  this.handleAdvancedSearch(
+                    mouseEvent,
+                    entity,
+                    query,
+                    searchSessionId,
+                    { resultsCount: 0, wasOnNoResultsScreen: true },
+                    isLoading,
+                  )
                 }
               />
             </NoResultsAdvancedSearchContainer>
@@ -196,10 +240,15 @@ export class JiraQuickSearchContainer extends React.Component<
               appPermission={appPermission}
               analyticsData={analyticsData}
               query={query}
-              showKeyboardLozenge={!isPreQuery && !keepPreQueryState}
-              showSearchIcon
               onClick={(mouseEvent, entity) =>
-                onAdvancedSearch(mouseEvent, entity, query, searchSessionId)
+                this.handleAdvancedSearch(
+                  mouseEvent,
+                  entity,
+                  query,
+                  searchSessionId,
+                  analyticsData,
+                  isLoading,
+                )
               }
             />
           </StickyFooter>
@@ -228,13 +277,21 @@ export class JiraQuickSearchContainer extends React.Component<
           mapSearchResultsToUIGroups(
             searchResults as JiraResultsMap,
             this.props.appPermission,
+            query,
           )
         }
         renderNoResult={() => (
           <NoResultsState
             query={query}
             onAdvancedSearch={(mouseEvent, entity) =>
-              onAdvancedSearch(mouseEvent, entity, query, searchSessionId)
+              this.handleAdvancedSearch(
+                mouseEvent,
+                entity,
+                query,
+                searchSessionId,
+                { resultsCount: 0, wasOnNoResultsScreen: true },
+                isLoading,
+              )
             }
           />
         )}
@@ -362,6 +419,7 @@ export class JiraQuickSearchContainer extends React.Component<
     query: string,
     sessionId: string,
     startTime: number,
+    queryVersion: number,
   ): Promise<ResultsWithTiming> => {
     const referrerId =
       this.props.referralContextIdentifiers &&
@@ -370,6 +428,7 @@ export class JiraQuickSearchContainer extends React.Component<
       query,
       { sessionId, referrerId },
       SCOPES,
+      queryVersion,
       JIRA_RESULT_LIMIT,
     );
 
@@ -446,6 +505,7 @@ export class JiraQuickSearchContainer extends React.Component<
       isSendSearchTermsEnabled,
       logger,
       enablePreQueryFromAggregator,
+      referralContextIdentifiers,
     } = this.props;
     const { selectedResultId } = this.state;
 
@@ -469,11 +529,14 @@ export class JiraQuickSearchContainer extends React.Component<
         }
         isSendSearchTermsEnabled={isSendSearchTermsEnabled}
         enablePreQueryFromAggregator={enablePreQueryFromAggregator}
+        referralContextIdentifiers={referralContextIdentifiers}
       />
     );
   }
 }
 
-export default injectIntl<Props>(
+const JiraQuickSearchContainerWithIntl = injectIntl<Props>(
   withAnalytics(JiraQuickSearchContainer, {}, {}),
 );
+
+export default withAnalyticsEvents()(JiraQuickSearchContainerWithIntl);