+++ packages/core/droplist/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/droplist
 
+## 9.0.1
+- [patch] [21854842b5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/21854842b5):
+
+  - Clean couple of TODO's that were already done
+
 ## 9.0.0
 - [major] [7c17b35107](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7c17b35107):